Category name should be lowercase and hyphenated
When you’re adding a shortcode like `[wpkb_list category="WUT CAtegoRy"
title="What Category Is This?!”]` the CSS being outputted based on the
category name should be in lowercase and hyphenated.

This fixes that, though this may not be the smartest way to do this.
Feel free to completely ignore this PR and fix this way smarter :)